Create Introduction to Programming – Mini Project
1. Create an encrypter in python based on the Ceaser cipher. It is a substitution cipher where each character of the original text is shifted to a certain number of characters in the alphabet. (encrypter and decrypter)
Write a function that would require 2 arguments – the input text to be encrypted and a key. For eg: Given the input text ‘hello’ and the key 3, the resulting encryted text would be ‘khoor’. Here you can see that every character in the string hello is shifted by 3 characters. ‘h’ has shifted to ‘k’, ‘e’ has shifted to ‘h’ and so on. If a key of 5 were used, the resulting string would be ‘mjqqt’. This function should be capable of ignoring any characters which are not alphabets. Th2 character ‘z’ entered b
y the user for a key of 3 would result in ‘c’.

Usage:<br>
**encrypt(‘hello world!’, 3)**<br>
**‘khoor zruog!’**

Similarly create decrypter which can decode the encryted text when provided the input text and key

Usage:<br>
**decrypt(‘khoor zruog!’, 3)**<br>
**‘hello world!’**

2.A Code to calculate the area of shapes. The program should is capable of calculating the area of a square, rectangle, triangle, and circle. The user is presented with options to select the shape. Based on which shape is closed by the user, the program asks for the appropriate input and prints the resulting area on the screen.