Add files via upload
Create an encrypter in python based on the Ceaser cipher. It is a substitution cipher where each character of the original text is shifted to a certain number of characters in the alphabet.

A Code to calculate the area of shapes. The program should is capable of calculating the area of a square, rectangle, triangle, and circle. The user is presented with options to select the shape. Based on which shape is closed by the user, the program asks for the appropriate input and prints the resulting area on the screen.